backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "job-matching-backend"
version = "0.1.0"
description = "AI-Based Job Profile Matching Backend"
authors = [{name = "Your Name"}]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "supabase>=2.3.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "PyPDF2>=3.0.1",
    "scikit-learn>=1.3.0",
    "nltk>=3.8.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.25.0",
    "apscheduler>=3.10.0",
    "slowapi>=0.1.9",
    "streamlit>=1.28.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]
requires-python = ">=3.9"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]